Move reused functions/components to separate files
